[docker-compose] Disable swap for highest priority services (#5892)
... by setting `deploy.resources.limits.memory` and `memswap_limit` to
99G. The idea of the very high limit is that we don't want to impose an
actual specific limit on memory usage by these services (so we're
setting it to a value that we will never reach), but, by setting these
values equal to each other, we disable the services from using swap at
all. For our highest priority services, this is basically what we want.

Note that we are _not_ applying this to `worker`, since we do want
`worker` to use a significant amount of swap (to leave physical RAM free
for the six more important services modified herein).