[docker-compose] Specify cpu_shares for all services [DEV-137] (#5899)
I used this grouping:

1. web services: 4096
2. databases: 2048
3. somewhat time-sensitive services: 1024
4. monitoring services: 512
5. one-off services without any need for performance: 256

In full, these are the priorities:

```
               nginx 4096
                 web 4096
            postgres 2048
           redis-app 2048
         redis-cache 2048
               clock 1024
              worker 1024
            cadvisor 512
             grafana 512
                loki 512
       node_exporter 512
          prometheus 512
       rails_metrics 512
              vector 512
             certbot 256
 initialize_database 256
    redis-overcommit 256
        s3_db_backup 256
```

I think this won't be relevant super often, since I think that the
`cpu_shares` are only relevant when the machine's available CPU core(s)
are being maxed out, which is not usually the case on the server
machine. However, CPU usage does usually get pegged to 100% during
deploys, so this might possibly help to keep the server more responsive
than it would otherwise be during deployments.