Disable @typescript-eslint/no-floating-promises rule
This is giving ~30 errors, and I don't really think it's worth much to
fix them (by marking the calls with `void` in front). These are
generally HTTP calls where I expect the response to succeed, and I don't
think it's worth the effort to handle the error. I think it's pretty
obvious from looking at the code that we aren't doing anything to handle
the promise's result, and I don't think `void` adds much/any clarity.