Fix specs
When inheriting from ApplicationController, the `name` of the controller
is set to `"AnonymousController"` by `rspec-rails`. This causes the
requests not to be recorded in Redis (since that is a #skip? condition),
which causes the tests to fail.

By inheriting from HomeController, instead, the controller name will not
trigger a #skip?.

Additionally, the `CONTROLLERS_NOT_TO_RECORD` constant has been deleted,
so I tweaked that spec to check for non-storage of request data in Redis
for an `uptime_robot` query param, instead.