[logs] Pass only 'log' prop to data renderers (#5924)
It seems sort of counterproductive / requires more typing / makes things
sort of confusing when we "destructure" a `log` into some of its
constituent properties (including `data_type`, `log_entries`, and
`data_label`), when there probably is no case where we would want to
pass a `log` object and then pass other `data_type`, `log_entries`,
and/or `data_label` that are not associated with that `log`. This gives
me the feeling that we should just use the single, unified `log` prop,
and destructure it as needed _within_ the data renderer components,
rather than destructuring as separate props to pass _to_ the component.
This changes the implementation from the destructured props approach to
the unified `log` prop approach.