Work around @typescript-eslint/no-misused-promises error
For some reason, this eslint error appears as a result of adding
`"vitest/globals"` to `"types": ["vite/client", "vitest/globals"]` in
`tsconfig.json`. ChatGPT says "The addition of vitest/globals to the
types array likely includes additional type definitions or global
augmentations that subtly affect TypeScript's behavior or type
inference. This might have caused TypeScript to infer the types in a way
that triggers the linter rule more aggressively."

https://chatgpt.com/share/6795b9b0-12d0-8007-ae7e-d2dab09a1bb7