Add vitest, @testing-library/vue, and jsdom (#5925)
`pnpm install -D vitest @testing-library/vue`

I used to have JavaScript tests, but I removed them in 5ba0c9f4 . I
don't remember exactly why. I think that the JavaScript tests were a
pain to write/maintain and added complexity to the bundling
configuration, etc. I have been pretty happy with the decision. I have
found that feature specs (written in Ruby, using Capybara) have provided
good coverage for the app's client-side functionality.

But, I think that now it is time to bring back some JavaScript tests.
For one thing, I think that the tooling is more developed, and with
vitest it is almost disturbingly easy to add a testing framework. For
another thing, there is a set of changes/bugfixes that I would like to
make, and for which I would like to have test coverage, but where I
think that adding to a feature spec might not be worthwhile. Our test
suite is starting to get sort of slow. Having the option to cover some
of the application's behavior with JavaScript specs rather than with
feature specs will be nice as an option to lessen our reliance on
feature specs (which are slow and can also be flaky).