[CompileJavaScript] Only print assets:upload_vite_assets on failure (#5927)
The output when this succeeds is pretty verbose (currently about 70
lines), and it's not very useful. This will keep the logs of test runs
on `main` more concise/readable.

Also, change a method name to reflect the important fact that the method
rescues all exceptions (not just ones inheriting from StandardError).