Explicitly sort text log entries by created_at [LOG-30] (#5929)
I think that we have been relying on the server-side ordering of the
text log entries, but that is insufficient if a new text log entry is
created that is backdated to be older than the most recent text log
entry. Prior to this change, that log entry would appear at the top of
the list. With this change, that log entry will enter the list of
entries in the proper chronological order.

Testing this was pretty brutal, and makes me question my decision to
reintroduce JavaScript testing in #5925 . The test actually seems to
take ridiculously long to run (between 2 and 3 seconds on my machine).
For now, though, I will continue ahead, and see if I find value in
having a JavaScript test harness available. Maybe it will not be so
painful, now that I have figured out some of the sticking points / added
some necessary and/or helpful supplemental libraries.