[groceries] Use icons for increment/decrement buttons [GROC-24] (#5930)
The font that was displaying these icons on my machine (and surely on at
least some other users' machines) does not align the horizontal bar of
the plus sign with the horizontal bar of the ndash. This created an
annoying visual effect where the ndash appeared to be off center (too
high).

Fortunately, the horizontal bars of the plus and minus icons in the
Tabler icons set _are_ aligned. By using those icons instead of plus and
ndash characters, we create a vertically aligned visual appearance for
these buttons.

See the "before" on the _right_ and after on the _left_ below:


![image](https://github.com/user-attachments/assets/5c6b05f2-ea53-4d30-a15c-69e085b0a99b)