[home] Shrink skill icons from 66px to 50px (mobile: 45px to 42px)
This change follows #5946 , which introduced a somewhat more bold color
palette for the skill icons. Probably partially because of that, I think
that the icons are too prominent. This change shrinks them a little bit,
so that they will be less prominent/distracting. The thing that we
probably most want users to focus on is the names of the skills, not the
pictures of them.