[emoji-picker] Define bootstrap data even if logged out [EMO-11]
When no user was logged in, `window.davidrunger.bootstrap` on the client
was undefined, which caused an error that totally broke the app (blank
white page).

This change fixes that by causing the assumption that
`window.davidrunger.bootstrap` will always be defined to (once again) be
true. However, we don't want to have to define a bootstrap JSON schema
for every single page, so we skip schema validation if the bootstrap
data only includes the basic/common/universal keys and if there is not
yet a JSON schema file for that page.

I believe that this bug was introduced here:
https://github.com/davidrunger/david_runger/pull/5045/files#r1732707807

If there is no logged in user, show a tip about logging in to access
boosting functionality.

Make `login` path exportable to the client with hack in
`config/routes.rb` to not alias that route when building JS routes.