Remove route aliasing for user and admin login paths [DEV-145] (#5958)
Our route "aliasing" in `config/routes.rb` for user and admin login
paths causes problems with `js-routes`.
(https://davidrunger.atlassian.net/browse/NEW-23 is to fix those issues
upstream.) But, I think that our route aliasing is not really needed /
worth it, in the first place. Therefore, this change removes those route
aliases, which allows us to use a `new_user_session_path` helper on the
frontend to refer to the `/login` path.