Stop truncating the requests table [DEV-142]
I think that the truncation of the `requests` table is mostly a legacy
from when the app was deployed to Heroku using a free Heroku database
that limited us to a total of 10,000 rows. Now that we are running the
app on a VPS with 25 GiB of storage, we have no fixed row limit and the
space considerations at this time are not significantly concerning. And
there is upside in being able to search over an indefinitely long
history of requests. So, this change removes the `requests` table from
the set of tables truncated by the `TruncateTables` worker.