Use propshaft for asset pipeline (and remove sprockets) [DEV-147] (#5973)
Sprockets is deprecated as of Rails 8. The new asset pipeline gem is
propshaft. To stay current, this change remove sprockets (and
sprockets-rails and sassc) and replaces it with propshaft.

Propshaft, however, does not perform any transformations (e.g. Sass to
CSS) -- and ActiveAdmin 3's stylesheets are written in CSS. Therefore,
in order to continue to use the ActiveAdmin stylesheets, we will now
compile them via Vite rather than via Sprockets. This excellent post
explained how to do that:
https://github.com/activeadmin/activeadmin/discussions/7947#discussioncomment-5867902
.

Unfortunately, there are lots of Sass deprecation warnings about the
ActiveAdmin Sass stylesheets. I guess that I will just ignore them, for
now. To that end, in this change, I am suppressing the output of the
admin assets compilation step in the `CompileJavaScript` test step.

(Side note: I think that ActiveAdmin 4 might be moving from Sass to pure
CSS?, in which case we might be able to remove the Vite compilation of
ActiveAdmin stylesheets.)

It was important to specify type="module" for the admin scripts so that
there aren't variable name conflicts between modules when multiple are
rendered on a page (e.g. how on the asset sizes admin page we include
both the active_admin script and the charts script).