[specs/feature] Write server log to file on spec failure [DEV-156] (#6003)
This should be helpful to debug failing or (especially) flaking specs.

I ran the test suite 3 times on `main` and then 3 times on this branch.
Results:

== main
1:29.49 total
1:23.02 total
1:20.45 total
Average: 1:24.3

== logs on failure
1:21.00 total
1:24.01 total
1:14.76 total
Average: 1:19.9

So, in those numbers, there's a 4.4 second speedup on this branch. I
can't think of how/why this branch would really provide a speedup vs
main, but the takeaway is that there does not appear to be a significant
slowdown, at least.