[admin/deploys] Make Git SHAs GitHub commit links [ADM-1]
This will make it more convenient / faster to see what a particular
deploy was deploying.

A single deploy could include multiple new commits, and so a compare
view from the previously deployed SHA to the SHA of each deploy in
question would probably be a better approach. However, that would be
somewhat difficult to do without an N + 1 query; at least, I don't think
it would be straightforward to do while simply leveraging  ActiveAdmin's
capacity to automatically construct most of the view. So, for now, for
simplicity, I am going to just ship this change. I think that hopefully
this 20% of the work will still deliver 80% of the possible benefits.

Also, while looking at the show page and creating a decorator, make the
main title prettier by defining a `to_s` method on the decorator.