Add CSV downloads of log data [LOG-27]
I am bothered that apparently element-plus doesn't have a "link that
looks just like a button" component, so I had to style it myself (which
is brittle and might start to look wrong if element-plus updates the
styling of their buttons, leaving my copy-pasted styling of the download
link behind). This is another data point adding to my feeling that it
might be best to move on from element-plus. But that will be for another
day.