Wrap 'MemoWise' in a generically named 'Memoization' class [DEV-112]
This will make it easier to swap out MemoWise for a different
memoization library, in the future (as we have done in the past
[once][1] and [twice][2] before). In general, I think it often makes
sense to "own" a light wrapper around services being used, whether that
is memoization or HTTP calls or a payment gateway.

Additionally, I don't really like the name `MemoWise`/`memo_wise`. It's
trying to be cute, but (not intending offense, and grateful for the
library) I personally find it a little annoying and cringe. I will
prefer just reading/writing `Memoization` and `memoize`.

[1]: https://github.com/davidrunger/david_runger/pull/2833
[2]: https://github.com/davidrunger/david_runger/pull/2834