Wrap 'MemoWise' in a generically named 'Memoization' class [DEV-112] (#6029)
This will make it easier to swap out MemoWise for a different
memoization library, in the future, if we ever want/need to, as we have
done in the past [once][1] and [twice][2] before. For this reason, in
general, I think it often makes sense to "own" a light wrapper around
services being used -- i.e. "[wrap your dependencies][3]" -- whether
that is memoization or HTTP calls or a payment gateway.

Additionally, I don't really like the name `MemoWise`/`memo_wise`. It's
trying to be cute, but (not intending offense, and grateful for the
library) I personally find it a little annoying and cringe. I will
prefer just reading/writing `Memoization` and `memoize`.

[1]: https://github.com/davidrunger/david_runger/pull/2833
[2]: https://github.com/davidrunger/david_runger/pull/2834
[3]: https://thoughtbot.com/blog/wrap-your-dependencies