Track scrolling with http/ky/fetch (not sendBeacon) [HOME-15]
Some ad blockers block `sendBeacon`, but I doubt any block `fetch`. We
use `sendBeacon` for external link clicks because those need to be sent
asynchronously (because otherwise the request might be cancelled as the
user navigates away). But we don't need that for scrolls. So, this
change uses `fetch` (via ky, via our http lib) to track scrolls, which
should be more reliable.

Also, disable `wrap_parameters` for JSON requests, which I find to be an
annoying (and, in the past, confusing) feature that I never asked for.
I'm surprised that I didn't notice sooner that it has been enabled.