Track scrolling with http/ky/fetch (not sendBeacon) [HOME-15] (#6045)
At least one ad blocker (uBlock Origin) blocks `sendBeacon`, but I doubt
that any block `fetch`. We use `sendBeacon` for external link clicks
because those need to be sent asynchronously (because otherwise the
request might be cancelled as the user navigates away), but we don't
really need that for scrolls. So, this change uses `fetch` (via ky, via
our http wrapper library) to track scrolls, which should be more
reliable.