Track length of CI build time steps [DEV-160]
Motivation: CI build times are getting kind of slow. Accumulating data
about the build times will be helpful to understand the problem(s) and
also to monitor the effectiveness of performance improvements that we
might attempt.

This change sort of undoes #720 , but whereas that PR removed
functionality that had leveraged the logs app and data model, this PR
adds a whole new, separate data model, controllers, and view.