[admin] Remove "Admin" menu [ADM-3]
Most of our models are turning out to be "Admin"-related. After all, I
guess that's the point of having an admin app...

Rather than cramming all of this stuff into an "Admin" menu, this change
removes the "Admin" menu and has everything now exist at the top level
of the ActiveAdmin menu.