[CiStepResult] Create CiStepResults in a single HTTP request [CSR-2] (#6057)
I started out using `runger_actions` for the service object to do this,
but it felt awkward and required remembering / looking up the particular
syntax for `runger_actions`. Instead, I switched to a PORO, which allows
me to use plain Ruby, which I know better than I know the specific
syntax/API for `runger_actions`. It felt nicer not to have to squeeze
what I am trying to do into the constraints of `runger_actions`. This
experience is leaning me somewhat toward thinking that `runger_actions`
might not be worthwhile.