Respond with JSON for 404 JSON requests [DEV-167]
See here:
https://github.com/davidrunger/david_runger/actions/runs/13222167275/job/36908608384#step:11:233

A request was made using JSON headers (via `Faraday.json_connection`)
for a route that doesn't exist, and Rails responded with the 404 HTML
page, rather than with JSON.

This change makes it so that we will now respond with JSON to such
requests.

(That this doesn't happen already/automatically (i.e. that I need to add
this `prioritize_json_format` method myself) sort of seems to me like a
bug in Rails?)