[pallets] Increase GitHub Actions parallelism from 2 to 4 [DEV-162]
GitHub Actions machines now have [4 vCPUs][1], and I think CI will run
fastest if we also set the pallets concurrency to 4.

To accomplish this, set a PALLETS_CONCURRENCY of 4 in the GitHub Actions
env. I think that it's appropriate to set this specifically there,
rather than relying on the default. However, I have also increased the
default / fallback value.

Also, print the pallets concurrency. This should help with awareness of
the concurrency level and will allow us to confirm that pallets is using
the expected concurrency level in GitHub Actions.

[1]: https://github.blog/news-insights/product-news/github-hosted-runners-double-the-power-for-open-source/