[pallets] Increase GitHub Actions concurrency from 2 to 3 [DEV-162] (#6062)
GitHub Actions machines now have [4 vCPUs][1], and I think CI will run
fastest if we also set the pallets concurrency to 3. I initially tried
4, but it seems that might be too much.
https://github.com/davidrunger/david_runger/actions/runs/13224908173/job/36914542355?pr=6062
Parallelism increased, but CPU time also significantly increased, so
that wall time (which is what we ultimately care about) is basically
right in line with the baseline that we have established with
concurrency of 2. Let's see if 3 is more of a sweet spot.

Moving from 4 to 3, we see that parallelism slightly decreased, but CPU
time increased even more, such that wall clock time decreased:


![image](https://github.com/user-attachments/assets/c72e9a08-a916-41e1-87a7-7501f660135b)

To accomplish this, set a PALLETS_CONCURRENCY of 3 in the GitHub Actions
`env`. I think that it's appropriate to set this specifically there,
rather than relying on the default. However, I have also increased the
default / fallback value.

Also, print the pallets concurrency. This should help with awareness of
the concurrency level and will allow us to confirm that pallets is using
the expected concurrency level in GitHub Actions.

[1]:
https://github.blog/news-insights/product-news/github-hosted-runners-double-the-power-for-open-source/