Split feature specs into three separate tasks [DEV-126]
This is quite rough/ugly, but I'm interested to see if it can
significantly improve test run time. I think that it might, in which
case I think that I might go forward with this, despite the significant
ugliness and to some extent risk of brittleness.

I guess that splitting into three feature spec groups makes sense
because currently we are running specs with a pallets concurrency of 3,
so no more than that many feature specs can run simultaneously, anyhow.