[logs/css] Remove 'display: inline-block' from 'p' tags
I have no idea why I had this (and I didn't look at the git history to
see if there are any clues). It's hard to think of a good reason to have
it, though. It can cause separate paragraphs to be rendered on the same
line, rather than separately, which I don't think that I ever want.
Therefore, I am removing it.