[spec/features/home] Check for PERCY_TOKEN (not PERCY_BUILD_ID)
... and bring sleep back down to 0.5 seconds.

`PERCY_BUILD_ID WAS NOT PRESENT!` was logged in CI
https://github.com/davidrunger/david_runger/actions/runs/13282554218/job/37083906146?pr=6107#step:11:428
which explains why the increasing sleeps didn't seem to be reducing the
flakiness. I think that PERCY_TOKEN is there, even if apparently
PERCY_BUILD_ID is now not there (presumably because of our switch from
`percy exec` to `percy exec:start`).