Split feature specs into two separate tasks [DEV-171] (#6107)
This is quite rough/ugly, but I'm interested to see if it can
significantly improve test run time. I think that it might, in which
case I think that I might go forward with this, despite the significant
ugliness and to some extent risk of brittleness.

I previously tried splitting them into _three_ separate tasks ( #6063 ),
but that didn't seem to help much.

However, I'm wondering if that's just because, between the server and
the RSpec process and the Chrome process, we were just trying to use too
much concurrency. But maybe two separate feature spec tasks running
concurrently will be faster than either one or three? This change tries
that.

Percy documentation about parallel builds related to some of these
changes:
https://www.browserstack.com/docs/percy/integrate/parallel-test-suites#parallel-tests-on-the-same-machine