Use previous prerendered page up to 5 minutes after deploy [DEV-114]
The amount of time that we want to cover is the amount of time from when
the relevant release task (added in this change) executes until a new
prerender is generated and uploaded to S3. It seems that this often
takes about 5 minutes, so we'll start with writing the cache value to
last for that long.