[prerenderable] Skip prerendered content if there's a prerender param (#6129)
Otherwise, due to #6127, we will serve the previously prerendered
content to the request to generate a new prerendered HTML page, so we
keep the old content alive indefinitely.

This param is added as a query param when we make the prerendering
request:
https://github.com/davidrunger/david_runger/blob/24ac6c4d4dac5e5de61b5b08316347fabd198cab/bin/prerender/#L23

Also, tweak the Vite Rails description, so that I'll be able to see
whether this works.