Change global element-plus import to component-local ones [DEV-176]
I think that this will result in some notably lighter page weights?

Another big benefit is that this gives us proper in-template TypeScript
support (e.g. type checking of the props) for the element-plus
components! Indeed, this surfaced several prop type violations, which
are corrected in this change.

Additionally, this will make it more feasible to start moving to a
different component library, if that is something that I ever want to
do, because, when the components are not registered globally, then we
can swap out element-plus for some other UI framework on an app-by-app
basis, rather than needing to do the whole swap all at once for all apps
(or incur extra JavaScript page weight, by (unnecessarily) loading
element-plus in addition to the new UI framework for apps that move to
the new framework).