Decrease file size allowances for JS and CSS in element-plus apps
Nice win! :D The logs JS and CSS didn't shrink that much, probably
because it (I'm guessing) was using most of the element-plus components
that we were importing anywhere, but the groceries and workout apps are
both seeing a significant reduction in both JS and CSS, which hopefully
will improve their performance (probably especially on mobile, where
both of those apps are most commonly used).