Change global element-plus import to component-local ones [DEV-176] (#6135)
One benefit is that this change results in a smaller JS and CSS page
weight for our element-plus-using apps, especially for groceries and
workout.

Another big benefit is that this gives us proper in-template TypeScript
support (e.g. type checking of the props) for the element-plus
components! Indeed, this surfaced several prop type violations, which
are corrected in this change.

Additionally, this will make it more feasible to start moving to a
different component library, if that is something that I ever want to
do, because, when the components are not registered globally, then we
can swap out element-plus for some other UI framework on an app-by-app
basis, rather than needing to do the whole swap all at once for all apps
(or incur extra JavaScript page weight, by (unnecessarily) loading
element-plus in addition to the new UI framework for apps that move to
the new framework).

I wonder if there's any possibility that this might even make build
times faster?, although I guess it also might make them slower. Hard to
say at this point (and maybe we'll never really be able to tell with any
confidence, and/or maybe there won't be much/any effect).