Wait for Percy to be running before starting feature specs (#6143)
We saw a Percy failure ([one][1], [two][2]) where Percy didn't receive
any snapshots. I think that this is because the Percy server wasn't yet
running when the Percy-snapshot-taking feature specs ran. I think that
this is because there was a new version of Chromium to download. In
general, I think that Percy will take a particularly long to start up if
there is a new Chromium version to download.

[1]:
https://percy.io/David-Runger/web/david_runger/builds/38736162/failed
[2]:
https://github.com/davidrunger/david_runger/actions/runs/13333515282/job/37243233305?pr=6142#step:11:166

To avoid this issue, this change adds a `WaitForPercyStart` test step
that uses `percy exec:ping` to make sure that Percy is running before
starting feature specs.