[ci] Increase Pallets concurrency from 4 to 5
Also, remove fallback when fetching the `PALLETS_CONCURRENCY` env var.
Anyone running pallets will now have to always provide that env var.
Motivation: I was getting a little sick of changing it in two places,
and I think arguably it's more clear to not have the fallback, so that
someone reading this code knows that the pallets concurrency is being
set by the env var being passed in rather than potentially coming from
the fallback value.