[vite] Set maxParallelFileOps for builds in CI (#6147)
I'm hoping that specifying these limits might reduce CPU contention and
make builds faster overall.

In CI, maxParallelFileOps will be 1 for the admin build, to hopefully
deprioritize it and cause less CPU contention (since it still seems to
finish reliably before the user JS build). For all other cases, we'll
set maxParallelFileOps to 4.

It's sort of hard to test whether this is having any effect, and I'm not
100% sure that it is. The admin JS builds don't seem much slower than
before. But I'll guess I'll merge it, in case it helps.