Split feature tests into 3 groups (not 2) (#6148)
I think that this will better make use of the 4 vCPUs that are on GitHub
Actions's machines. I tried this before (#6063), and it didn't seem to
work, but I didn't have some things in place then that I have now (like
more robust and efficient Percy setup for parallel runs, and also the
Gantt charts, which are super helpful to understand what's going on), so
I think it might work well this time around, i.e. I think it might
decrease overall run time.