[JsonSchemaValidator] Automatically open the schema file in editor (#6151)
This will make it a little easier to paste the JSON Schema into the
appropriate file (since it will now be opened automatically in the
developer's editor). We were already creating and opening the JSON
Schema file if it didn't exist, but this change makes it so that we will
also open the file even if it does already exist, for easier editing (or
pasting in an updated version of the JSON Schema generated by the
converter tool, which will also be opened automatically).