Don't modify globally shared filesystem in CSV download spec [LOG-36] (#6155)
In the spec being modified herein, the filesystem at
`Capybara.save_path` was acting essentially as shared global state among
all specs. I think that there was at least a theoretical risk that the
spec could conflict with another spec that also modified the state of
files in `Capybara.save_path` .

To fix, before running feature specs, we will set `Capybara.save_path`
to a unique, randomly generated, temporary directory, in order to
essentially carve out a space in the file system that is local only to
this suite of feature specs. This change also provides a
`downloaded_file_path` helper method to then access files downloaded to
that directory.

Note that this solution doesn't 100% solve the issue of global state
modification; we've just moved the problem from modifying a global
directory in the filesystem to instead modifying the global
`Capybara.save_path`. Still, this is an improvement, because the file
system is shared by all processes on the same machine, whereas
`Capybara.save_path` is only shared by simultaneously executing Ruby
threads in the same process. So, the scope of global modification is at
least made smaller, and since I don't think that we run RSpec tests in
parallel threads within the same process, that shouldn't be a risk to
us.